Dev (#4)
* Delegate method is unified now. It returns ExpyState instead of ExpyActionType.
ExpyTableViewCell protocol is added. If a cell conforms to it, it gets state change updates in the changeState method.
Method named getActionType is removed.
ExpyState with didExpand and didColapse works more accurate now.(They were being called before the animation completely finished.)
First cell which is used to generate an expandable header is not being reloaded anymore during the expanding or collapsing animation. Because, this was preventing implementation of independent animations in it. (For example rotating the arrow in header cell.)
Improving examples.

* When an expandable cell tapped, its userInteractionEnabled parameter becomes false until the animation is completed.

* Adding cellReuseStatus to changeState method of ExpyTableViewHeaderCell

* Fixes #3